% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar2.R
\name{ar2}
\alias{ar2}
\title{Obtain the adjusted R-Squared Value from a Linear Model}
\usage{
ar2(object, data = NULL)
}
\arguments{
\item{object}{An R object that is a `formula` or contains the results of the `lm` function.}

\item{data}{A data frame when the object is a formula.}
}
\description{
Obtain the adjusted R-Squared Value from a Linear Model
}
