% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{Extract the regression coefficients}
\usage{
b(object, index = NULL)
}
\arguments{
\item{object}{An R object that contains the results of the `lm` function.}

\item{index}{Index indicating which coefficients to obtain.}
}
\description{
The regression coefficients demonstrates how a set of predictor
variables will affect the outcome of interest
}
