% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{cat_stats}
\alias{cat_stats}
\title{Obtain Frequencies and Proportions for a Categorical Variable}
\usage{
cat_stats(x)
}
\arguments{
\item{x}{Vector}
}
\description{
Obtain Frequencies and Proportions for a Categorical Variable
}
