% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csucistats-package.R
\docType{package}
\name{csucistats-package}
\alias{csucistats}
\alias{csucistats-package}
\title{csucistats: CSU Channel Islands R Tools}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An R package containing functions for statistics courses at CSUCI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.inqs.info/csucistats/}
}

}
\keyword{internal}
