% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documents.R
\name{new_presentation}
\alias{new_presentation}
\title{Generate a QMD file to create a presentation template for class. Files will be saved in it's own folder.}
\usage{
new_presentation(
  hw_name,
  name = "Name Here",
  title = "Title Here",
  date = "`r format(Sys.time(),'\%m-\%d-\%Y')`",
  dir_loc = getwd(),
  data = FALSE
)
}
\arguments{
\item{hw_name}{Name of folder to store all important files for the homework assignment. This argument will remove unwanted punctuations and spaces.}

\item{name}{Your name.}

\item{title}{Optional title to give to assignment.}

\item{date}{Optional Date to give to the document.}

\item{dir_loc}{Optional location to place the folder and hw template contents.}

\item{data}{If TRUE, a data folder will be constructed in hw template folder.}
}
\description{
Generate a QMD file to create a presentation template for class. Files will be saved in it's own folder.
}
