% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{num_stats}
\alias{num_stats}
\title{Obtain Numerical Statistics for a Continuous Variable}
\usage{
num_stats(x)
}
\arguments{
\item{x}{A numerical or integer vector.}
}
\description{
Obtain Numerical Statistics for a Continuous Variable
}
