% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/press.R
\name{press}
\alias{press}
\title{Repeat a task numerous times}
\usage{
press(expr, sim)
}
\arguments{
\item{expr}{Expression to be evaluated multiple times.}

\item{sim}{The number of times to evaluate an the expression.}
}
\description{
Repeat a task numerous times
}
