% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props_df.R
\name{props_df}
\alias{props_df}
\title{props Computing proportions}
\usage{
props_df(df, x, y, yval, diff = FALSE)
}
\arguments{
\item{df}{data frame}

\item{x}{grouping variable}

\item{y}{outcome of interest}

\item{yval}{category of interest from outcome of interest}

\item{diff}{Obtaine the difference in proportions. Default is FALSE.}
}
\description{
Compute the proportions of observing a value from a 2 by 2 continguency table
}
