% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_df.R
\name{resid_df}
\alias{resid_df}
\title{Extract Residuals and Influential Measures from and `lm` object}
\usage{
resid_df(object)
}
\arguments{
\item{object}{An R object that contains the results of the `lm` function.}
}
\description{
Extract Residuals and Influential Measures from and `lm` object
}
