% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse}
\alias{sse}
\title{Compute the sum of error squared for an R object}
\usage{
sse(object)
}
\arguments{
\item{object}{An R object that contains the results of the `lm` function.}
}
\description{
Compute the sum of error squared for an R object
}
