% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorn.R
\name{unicorn}
\alias{unicorn}
\title{Creates a Unicorn Data Set from simulated models.}
\usage{
unicorn(n)
}
\arguments{
\item{n}{Number of unicorns used}
}
\description{
Creates a Unicorn Data Set from simulated models.
}
