% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Obtained the R-Squared Value from a Linear Model}
\usage{
r2(object)
}
\arguments{
\item{object}{An R object that contains the results of the `lm` function.}
}
\description{
Obtained the R-Squared Value from a Linear Model
}
